+++ test/lib/getJsdocAstTag.test.js
@@ -1,36 +1,41 @@
 'use strict';
 
-const { deepStrictEqual, strictEqual } = require('assert');
-const doctrine = require('doctrine');
+const { strictEqual } = require('assert');
+const { resolve } = require('path');
+const commentParser = require('comment-parser');
+const snapshot = require('snapshot-assertion');
 const getJsdocAstTag = require('../../lib/getJsdocAstTag');
 
 module.exports = (tests) => {
-  tests.add('`getJsdocAstTag` with a tag.', () => {
-    deepStrictEqual(getJsdocAstTag(doctrine.parse('@name a').tags, 'name'), {
-      title: 'name',
-      description: null,
-      name: 'a',
-    });
+  tests.add('`getJsdocAstTag` with a tag.', async () => {
+    const [jsdocAst] = commentParser('/** @name a */');
+
+    await snapshot(
+      JSON.stringify(getJsdocAstTag(jsdocAst.tags, 'name'), null, 2),
+      resolve(__dirname, '../snapshots', 'getJsdocAstTag', 'with-a-tag.json')
+    );
   });
 
-  tests.add('`getJsdocAstTag` with a tag override.', () => {
-    deepStrictEqual(
-      getJsdocAstTag(
-        doctrine.parse(
-          `@name a
-@name b`
-        ).tags,
-        'name'
-      ),
-      {
-        title: 'name',
-        description: null,
-        name: 'b',
-      }
+  tests.add('`getJsdocAstTag` with a tag override.', async () => {
+    const [jsdocAst] = commentParser(`/**
+ * @name a
+ * @name b
+ */`);
+
+    await snapshot(
+      JSON.stringify(getJsdocAstTag(jsdocAst.tags, 'name'), null, 2),
+      resolve(
+        __dirname,
+        '../snapshots',
+        'getJsdocAstTag',
+        'with-a-tag-override.json'
+      )
     );
   });
 
   tests.add('`getJsdocAstTag` with no tag.', () => {
-    strictEqual(getJsdocAstTag(doctrine.parse('').tags, 'name'), undefined);
+    const [jsdocAst] = commentParser('/** Description. */');
+
+    strictEqual(getJsdocAstTag(jsdocAst.tags, 'name'), undefined);
   });
 };