+++ test/lib/getJsdocAstTags.test.js
@@ -1,40 +1,30 @@
 'use strict';
 
-const { deepStrictEqual, strictEqual } = require('assert');
-const doctrine = require('doctrine');
+const { strictEqual } = require('assert');
+const { resolve } = require('path');
+const commentParser = require('comment-parser');
+const snapshot = require('snapshot-assertion');
 const getJsdocAstTags = require('../../lib/getJsdocAstTags');
 
 module.exports = (tests) => {
-  tests.add('`getJsdocAstTags` with @param.', () => {
-    deepStrictEqual(
-      getJsdocAstTags(
-        doctrine.parse(
-          `Description.
-@kind function
-@name a
-@param {string} a Description.
-@param {string} b Description.`
-        ).tags,
-        'param'
-      ),
-      [
-        {
-          title: 'param',
-          description: 'Description.',
-          type: { type: 'NameExpression', name: 'string' },
-          name: 'a',
-        },
-        {
-          title: 'param',
-          description: 'Description.',
-          type: { type: 'NameExpression', name: 'string' },
-          name: 'b',
-        },
-      ]
+  tests.add('`getJsdocAstTags` with @param.', async () => {
+    const [jsdocAst] = commentParser(`/**
+ * Description.
+ * @kind function
+ * @name a
+ * @param {string} a Description.
+ * @param {string} b Description.
+ */`);
+
+    await snapshot(
+      JSON.stringify(getJsdocAstTags(jsdocAst.tags, 'param'), null, 2),
+      resolve(__dirname, '../snapshots', 'getJsdocAstTags', 'with-@param.json')
     );
   });
 
   tests.add('`getJsdocAstTags` with no tags.', () => {
-    strictEqual(getJsdocAstTags(doctrine.parse('').tags, 'param'), undefined);
+    const [jsdocAst] = commentParser('/** Description. */');
+
+    strictEqual(getJsdocAstTags(jsdocAst.tags, 'param'), undefined);
   });
 };