+++ test/lib/typeJsdocStringToJsdocAst.test.js
@@ -4,25 +4,86 @@ const { deepStrictEqual, throws } = require('assert');
 const typeJsdocStringToJsdocAst = require('../../lib/typeJsdocStringToJsdocAst');
 
 module.exports = (tests) => {
-  tests.add('`typeJsdocStringToJsdocAst` with a parameter type.', () => {
-    deepStrictEqual(typeJsdocStringToJsdocAst('...*', true), {
-      type: 'RestType',
-      expression: {
-        type: 'AllLiteral',
-      },
-    });
-  });
+  tests.add(
+    '`typeJsdocStringToJsdocAst` with non-paramter type, non-optional.',
+    () => {
+      deepStrictEqual(typeJsdocStringToJsdocAst({ type: 'a' }), {
+        type: 'NameExpression',
+        name: 'a',
+      });
+    }
+  );
 
-  tests.add('`typeJsdocStringToJsdocAst` with a non-parameter type.', () => {
-    deepStrictEqual(typeJsdocStringToJsdocAst('a'), {
-      type: 'NameExpression',
-      name: 'a',
-    });
-  });
+  tests.add(
+    '`typeJsdocStringToJsdocAst` with non-parameter type, optional via option.',
+    () => {
+      deepStrictEqual(
+        typeJsdocStringToJsdocAst({ type: 'string', optional: true }),
+        {
+          type: 'OptionalType',
+          expression: {
+            type: 'NameExpression',
+            name: 'string',
+          },
+        }
+      );
+    }
+  );
 
-  tests.add('`typeJsdocStringToJsdocAst` with an invalid type.', () => {
+  tests.add(
+    '`typeJsdocStringToJsdocAst` with parameter type, non-optional.',
+    () => {
+      deepStrictEqual(
+        typeJsdocStringToJsdocAst({ type: '...*', parameter: true }),
+        {
+          type: 'RestType',
+          expression: {
+            type: 'AllLiteral',
+          },
+        }
+      );
+    }
+  );
+
+  tests.add(
+    '`typeJsdocStringToJsdocAst` with parameter type, optional via type string.',
+    () => {
+      deepStrictEqual(
+        typeJsdocStringToJsdocAst({ type: 'string=', parameter: true }),
+        {
+          type: 'OptionalType',
+          expression: {
+            type: 'NameExpression',
+            name: 'string',
+          },
+        }
+      );
+    }
+  );
+
+  tests.add(
+    '`typeJsdocStringToJsdocAst` with parameter type, optional via type string and option.',
+    () => {
+      deepStrictEqual(
+        typeJsdocStringToJsdocAst({
+          type: 'string=',
+          parameter: true,
+          optional: true,
+        }),
+        {
+          type: 'OptionalType',
+          expression: {
+            type: 'NameExpression',
+            name: 'string',
+          },
+        }
+      );
+    }
+  );
+
+  tests.add('`typeJsdocStringToJsdocAst` with invalid type.', () => {
     throws(() => {
-      typeJsdocStringToJsdocAst('**');
+      typeJsdocStringToJsdocAst({ type: '**' });
     }, new Error('Invalid JSDoc type “**”.'));
   });
 };